Refactor how vectors of data is stored
Previously we had special Data types for a vector of points and line segments.

This PR simplifies this to instead use the type-erased `DataVec`.

This makes the code simpler, more general, and more dynamic -
but it moves type checking into a later staged.

It's very likely we want to refactor and generalize this further,
for instance to support tensors.