Improve time range queries: also select latest before range (#98)
Conceptually, each piece of data on a certain path lasts until the next
piece of data is logged. That is, the times are the _start times_.

Hence, when you select a time range in the UI you want to see all the
data logged in that range, plus the latest data logged just before.

For instance: if you log a "scene" mesh (e.g. a street) at time=0
(or, soon, timeless) then you want that scene to always be visible.

Related to timeless data, PRO-97