Timeless data (#96)
This adds a timeless=False parameter to all Python logging functions. Timeless objects belong to all timelines (existing ones, and ones not yet created) and are shown leftmost in the time panel in the viewer. This is useful for object that aren't part of normal data capture, but set the scene for how they are shown. For instance, if you are logging cars on a street, perhaps you want to always show a street mesh as part of the scenery, and for that it makes sense for that data to be timeless.

set_space_up is also logged as timeless, meaning you can set it before adding any other timelines, and it will still wrok as expected.

This adds a special case of an empty TimePoint to mean "timeless data". This data is added to all existing timelines, and to all future timelines (yet to have been created).

The way this is implemented is that timeless data is added to each timeline at TimeInt::BEGINNING which is simply a time far into the past, and shown as "-∞". This makes the timeless data show up in each query (as they should) without any special cases in the queries. It also allows users to overwrite timeless data with either new timeless data (overwriting the previous one) or overwriting it with new timeful data.

The way this is implemented is that LogDb detect and handle this special case. It is a little bit hairy.