Add ability to run re_viewer on wgpu
This adds the features 'glow' and 'wgpu' to re_viewer.

If `glow` is enabled (default), we use `glow` for eframe
and `three-d` for rendering.

If `wgpu` is enabled and `glow` is not, we use `wgpu` for eframe,
and the 3D scenes are all disabled (because we don't have a wgpu
renderer yet).