Introduce API for logging a class_descriptions and segmentation_image (#187)
Introduce a new API for logging a set of ClassDescriptions which can be used to colorize 2d images of type uint8 and uin16

The new API takes a list of ClassDescriptions(id=int, label=str, color=(r,g,b)). If color is left unpopulated they will be generated automatically. Colors are "stable" determined by the id / value.

Your image must be logged with a reference to the path of the segmentation map.