Special case for Batch storage of Sequential Indices (#219)
In order to improve the storage efficiency of dense sequential batches (such as pointclouds) we special case the handling of sequential indices.

This replaces instances of Vec used in the Batch APIs with a new enum: BatchIndex
```
pub enum BatchIndex {
    /// Many batches use implicit sequential-indexing
    SequentialIndex(usize),

    /// Some batches want to provide explicit indices
    FullIndex(Vec<Index>),
}
```

A SequentialIndex lets us indicate a monotonically increasing sequence of indices up a a max without needing to re-transmit them.

We then implement two alternative storage models for Batch
```
pub enum Batch<T> {
    SequentialBatch(Vec<T>, ArcIndexHashes),
    IndexedBatch(IntMap<IndexHash, T>),
}
```

In SequentialBatch we store a dense vector and use a separate Arc for the collection of IndexHashes as necessary. IndexedBatch continues to store the elements within a per-batch map (with the overhead that entails).