Filter out wgpu spam in web console log (#292)
Closes PRO-255

This also seems to mitigate PRO-256. At least I fail to replicate it
after this. If the problem with `PRO-256` is that it _sometimes_
crashes when there is any logging, then this PR will make the
bug much less likely, but not solve it. So we still need to figure out
PRO-256.