[re_renderer] Introduce `DrawPhase`es for simpl(istic) multi pass rendering Or rather as a first and super easy step towards it. Also takes the first step towards supporting transparency, see #702 Replaces previous sorting key!
Very simple idea: Every `Renderer` is invoked with every `DrawPhase` it registered for.
`DrawPhase` have a strict order and implicitly describe the `wgpu::RenderPass` that is passed for drawing.

For now and forseeable future, these passes are managed by the ViewBuilder (unchanged!).
At some point we will probably make `DrawPhase` a more complex object, but this is not today!