Optimize the depth-cloud shader when depth=0
depth=0 is a degenerate case that causes a 1440x1920 depth-map
from https://github.com/rerun-io/rerun/pull/1538 take 75ms (!) on my
M1 MacBook Pro.

With this fix this goes down to 15 ms

However, if we zoom out a lot, so that the depth point cloud covers
a very small part of the screen, the slowness returns.

I suspect this is a GPU binning problem of some sort, where too many
points in too small of a screen area causes performence issues.