Initial Space View trait & port of text space views to the new Space View trait system (#2281)
<!--
Open the PR up as a draft until you feel it is ready for a proper
review.

Do not make PR:s from your own `main` branch, as that makes it difficult
for reviewers to add their own fixes.

Add any improvements to the branch as new commits to make it easier for
reviewers to follow the progress. All commits will be squashed to a
single commit once the PR is merged into `main`.

Make sure you mention any issues that this PR closes in the description,
as well as any other related issues.

To get an auto-generated PR description you can put "copilot:summary" or
"copilot:walkthrough" anywhere.
-->

### What

Main pieces of:
* #2249 
* #1873 

Introduces a new framework for space view classes that will eventually
replace `ViewCategory` (right now the systems live side by side,
creating some oddities).
Ports text & text-box space views to this new system.

### Why

This paves the way for more structured space views, more streamlined
blueprint configuration and user defined space views.
In fact, this PR already enables user defined space views, but does not
yet expose a way to add them to the viewport (this can be done with some
small hacks to the space view adding code though and works very well!)

### Future work / discussion

There is still a lot of open question on the space view trait and it
will require changes as we move the other space views over to it.

Most notably, archetypes are defined by have no effect yet!
Overall, scene definition can be regarded as experimental at this point.
I chose to have a hard separation of `SceneElement` (the successor of
`ScenePart`) in the hope of providing a more powerful framework and an
easy hook for future parallelization. We'll need to see how this pans
out though!

Unlike planned, the definition of the space view class trait (as well as
implementation utilities) are not in a separate crate, but part of the
viewer context since global access to the space view type registry
proved very valuable.
We could still separate some parts of it out if desirable, but this
seems not very important at the moment.

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)

<!-- This line will get updated when the PR build summary job finishes.
-->
PR Build Summary: https://build.rerun.io/pr/2281