Expose type erased draw data that can be consumed directly
Split out of ongoing scene traitification refactor: The `SceneElement` trait should be able to emit draw data directly and this smaller refactor allows that.
This happens to avoid a lot of copy operations we had and makes use of DrawData safer as it enforces the "single frame" assumption a bit better.