add manifest + readmes to examples (#2309)
<!--
Open the PR up as a draft until you feel it is ready for a proper
review.

Do not make PR:s from your own `main` branch, as that makes it difficult
for reviewers to add their own fixes.

Add any improvements to the branch as new commits to make it easier for
reviewers to follow the progress. All commits will be squashed to a
single commit once the PR is merged into `main`.

Make sure you mention any issues that this PR closes in the description,
as well as any other related issues.

To get an auto-generated PR description you can put "copilot:summary" or
"copilot:walkthrough" anywhere.
-->

### What

Closes #2268 
Part of #2267 

Adding a `manifest.yml` file which is used by `landing` to render
examples in the documentation, and a `README.md` to each example.

GitHub renders frontmatter as a table, I'm not sure if we want to go
with that anymore. It's also possible to do something else, like a
comment at the top with a custom format, which wouldn't be displayed,
but would still be present in the file which `landing` downloads, and
could then be used to store the same metadata.

### Checklist
* [ ] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [ ] I've included a screenshot or gif (if applicable)

<!-- This line will get updated when the PR build summary job finishes.
-->
PR Build Summary: https://build.rerun.io/pr/2309

<!-- pr-link-docs:start -->
Docs preview: https://rerun.io/preview/a05e331/docs
<!-- pr-link-docs:end -->

---------

Co-authored-by: Nikolaus West <niko@grapic.co>