Python codegen: big cleaning and paving the way towards transforms (#2603)
### What

**Note**: as usual, review commit-by-commit (the bulk of the changes are
in the first commit). This time, reviewers might want to spend some time
on the auto-generated code, though, as this is a major focus of this PR.

This PR has the following aims:
- Enforce and exploit the decision that a component must be a struct of
one, and exactly one, field, which may be a datatype (in which case the
component is known as "delegating", as its implementation delegates to
the datatype) or a native type (e.g. float, int, etc.).
- Significantly polish the Python SDK, with the goal of a cleaner, fully
typed user-facing API, and a streamlined implementation. Part of this is
achieved using `attrs` (see below).
- Improve the `from_similar()` hook and introduce several others. The
proposed mechanisms consist of storing the so-called (hand-coded)
overrides in the `_overrides` sub-packages in each of the `archetypes`,
`components`, and `datatypes` packages. These overrides are discovered
by the code generator by "parsing" (well, "brute-force extracting
symbols" from) `_overrides/__init__.py` [1]. The following hooks are
currently implemented:
- `xxx_native_to_pa_array()` (where `xxx` is the lower-cased name of the
corresponding object): this is the primary way to flexibly convert
native input into the Arrow array form of the component.
- `xxx_yyy_converter` (`yyy` is a field name): optional converter
function a given field of a native objects (e.g. `Color`). This enables
the object to flexibly accept native input for the given field.
- `xxx_as_array`: optional `__array__()` magic function for native
objects (e.g. `Point2D`). This function enable the object to be
interpreted as array data by Numpy, which can considerably simplify the
implementation of, e.g., `xxx_native_to_pa_array` (`Point2D` is a good
example).

This PR also introduces the following changes:
- So far, the archetypes and `log_any()` were pulled into the `rr`
namespace. This PR now pulls all component- and datatype-related symbols
in `rr.cmp` and `rr.dt`, respectively. This is intended as a temporary
solution, to avoid name clashes with the legacy API [2].
- The [`attrs`](https://www.attrs.org/en/stable/) dependency is
introduced as it allows much more flexibility than the standard library
`dataclass`, in particular with the `converter` feature which I use
extensively. Introducing new deps is always a tough call. In the case of
`attrs`, it's most probably safe as this is one of the most common and
well maintained package of the ecosystem (110+M dl/month).
- Native objects automatically get `__str__()`, `__int__()`, or
`__float__()` functions when they contain a single field of the related
type.
- The `Point2D` component now delegate to a new `Point2D` datatype,
defined as a struct of two floats. This is done such as to avoid
breaking compatibility with the current viewer. The long-term goal is to
use the `Vec2D` datatype instead.
- The CI now uses Python 3.11 for linting (and linting only, i.e. not
testing, etc.)

Addressing union types is a non-goal of this PR. This will be part of
the next PR (transforms). As such `QuotedObject::from_union()` is most
likely broken.

TODO/TBD:
- PyCharm's typechecker isn't happy with the Archetype, with is
irritating. This could apparently be addressed by using `xxxArrayLike`
instead of `xxxArray` annotations for archetype's fields. I'm curious
how VSCode fares, though.
- Maybe passing `None` as args to the archetype should set `None`
instead of empty arrays (they aren't logged either way).
- Maybe the notion of delegating vs. non-delegating component should be
promoted to the langauge-agnostic code?


[1] I actually prototyped and tested an approach based on PyO3. It sorta
worked, until I tried to import stuff (Numpy, ...) for the overrides.
The codegen would somehow be given access to a venv with the required
deps, etc. Hell no.
[2] Name clashes also exist internally to the new API, e.g.
`components.Point2D` vs. `datatypes.Point2D`. This will have to be
addressed before we can pull "everything" into the `rr` namespace.

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested [demo.rerun.io](https://demo.rerun.io/pr/2603) (if
applicable)

- [PR Build Summary](https://build.rerun.io/pr/2603)
- [Docs
preview](https://rerun.io/preview/pr%3Aantoine%2Fhope-python-cleanup-new/docs)
- [Examples
preview](https://rerun.io/preview/pr%3Aantoine%2Fhope-python-cleanup-new/examples)

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>