Fix blueprint heuristic updates not propagating through the tree unless there were manual changes The issue was easiest to observe in the transform demo which relies completely on heuristics to display anything (all transform arrows have a configurable size): On startup or after reset nothing is visible until something was selected. This happened because `class.prepare_ui` sets the size (via heuristic) on the blueprint but `class.ui` would see the propagated results only by next frame. At this point we discard the entire blueprint unless there are manual changes. It seems changing the selection in the blueprint alone is enough to trigger such a change.
To fix this I'm now simply applying the heuristic changes in the same frame.

However, there seems to be a deeper underlying issue where until the first selection (or other change) is done we're discarding the blueprint every frame - once something is selected `sync_space_view` will overwrite the previous snapshot and heuristic results will be applied. This might be intended but seems rather odd. Expecation is that a lot of this will go away once we reworked blueprints/properties as planned. cc: @jleibs