Introduce versioned `EntityPath` & refactor mesh/tensor caching (#3230)
**Commit by commit**

What the title says.

The most important contribution of this PR is the new
`VersionedInstancePath[Hash]` types, which further specializes an
`InstancePath[Hash]` (itself a specialization of `EntityPath`) with
extra versioning information in the form of a `RowId` from the store.
I.e. a `VersionedInstancePath` gives you a specific instance of an
entity at a specific path at a specific time and row in the store, which
is as specific as it gets.

This allows us to completely remove `MeshId` & `TensorId`, and paves the
way for generic frame-based caching.