Update all uses of `actions/checkout` to use explicit `ref` (#3322)
<!--
Open the PR up as a draft until you feel it is ready for a proper
review.

Do not make PR:s from your own `main` branch, as that makes it difficult
for reviewers to add their own fixes.

Add any improvements to the branch as new commits to make it easier for
reviewers to follow the progress. All commits will be squashed to a
single commit once the PR is merged into `main`.

Make sure you mention any issues that this PR closes in the description,
as well as any other related issues.

To get an auto-generated PR description you can put "copilot:summary" or
"copilot:walkthrough" anywhere.
-->

### What

`actions/checkout` uses `GITHUB_SHA` by default for the `ref` which it
checks out. This sometimes leads to confusing issues, because the
`pull_request` commit actually operates not on the latest commit of a
pull request, but on the latest _merge_ commit of a pull request. This
causes a lot of confusion and even actual bugs.

In this PR, I'm setting all the `ref` values explicitly to whatever is
available:
- Some reusable workflows already require a commit, .e.g
`reusable_build_and_publish_web` has a `release-commit` input
- If the workflow can be triggered from a PR, then we use
`github.event.pull_request.head.ref`, which points to the HEAD of the PR
branch.
- In all other cases we fall back to `github.sha`. For example a `push`
to `main` will still use `github.sha`, which is fine.

**There are caveats**. With these changes, we no longer test PRs on the
merge commit. That means if your PR is out-of-date with the target
branch (e.g. `main`), and someone pushes to `main`, you are less likely
to find out about any potential incompatibility if it doesn't cause a
merge conflict. This means we need to be careful about merging without
being up-to-date with `main`. Worst case scenario it will result in more
failures on `main` which we'll be notified about on Slack.

The main benefit of this is less confusion about which commit we're on
when viewing the web demo, which is an essential part of our overall
workflow at this point. Once we get over the `source_hash` situation and
have the ability to turn on merge queue, we'll once again have much more
"main safety".

Before: 

![image](https://github.com/rerun-io/rerun/assets/1665677/df572ccf-1a4b-44f3-ac1d-99d4f224ec1f)


After:

![image](https://github.com/rerun-io/rerun/assets/1665677/ee220ba0-ef62-4f85-a58a-43b8a5d8b146)

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested [demo.rerun.io](https://demo.rerun.io/pr/3322) (if
applicable)

- [PR Build Summary](https://build.rerun.io/pr/3322)
- [Docs
preview](https://rerun.io/preview/99f0bc92ea87cfc1868d11173b632ba6d8801e72/docs)
<!--DOCS-PREVIEW-->
- [Examples
preview](https://rerun.io/preview/99f0bc92ea87cfc1868d11173b632ba6d8801e72/examples)
<!--EXAMPLES-PREVIEW-->
- [Recent benchmark results](https://ref.rerun.io/dev/bench/)
- [Wasm size tracking](https://ref.rerun.io/dev/sizes/)