Make entity path parsing infallible by default
Closes https://github.com/rerun-io/rerun/issues/3393

For the Python and Rust SDK, any string is now a valid entity path,
though perhaps not the one you where expecting. A warning is printed
if the format is not the normalized format.

For C/C++ I opted to still use the strict parsing for now.
We can revisit that later.