Fixes no space views showing up for later logged entities (#3454)
### What

* Fixes #3413

Currently, if any significant user change occurs to the blueprint we no
longer add any new space views. For this we relied on detecting
arbitrary changes to the tile tree. This causes a lot of false positive
change detections. Setting `has_been_user_edited` is more explicit now
again. Major drawback is that this easily doesn't detect all possible
changes. For instance, with the fix as is it doesn't detect manual
relayouting as a change - which arguably isn't enough of a change to
warrant not adding new space views. Eventually, we'll have to put the
long planned ui for auto-adding space views in place, but this fix will
have to do for now


Tested with
```
rm -rf ~/Library/Application\ Support/rerun/blueprints
python ./examples/python/plots/main.py      
```

![image](https://github.com/rerun-io/rerun/assets/1220815/4d16b02a-c2fe-49df-a54a-99d6d40d1029)