Use `macos-latest-large` runners when building wheels and rerun-cli
In our latest release, building rerun-cli for mac took 35min on arm
and 32min on x64. Windows took 14 min and ubunut 8min.

Building the wheels for mac took 25min on arm and 22min on x64.
Windows took 13 min, and ubuntu 7min.

`macos-latest-large` is a 12-core x64 machine.
There is also `macos-latest-xlarge` we can try, which is ARM based.

That's at least if I understand this blog post correctly:
https://github.blog/2023-10-02-introducing-the-new-apple-silicon-powered-m1-macos-larger-runner-for-github-actions/

That is the only docs I can find about this, at all.