Formalize behaviour of our many time histograms in a test suite (#4215)
We maintain three different kinds of time histograms as part of our
`EntityDb`:
- `times_per_timeline`: instantiated once per `EntityDb`, this keeps
track of the number of distinct times per timeline that have been logged
to, across all entities and components.
- `prefix_times`: instantiated once per `EntityTree` (and thus once per
entity path), this keeps track of the number of raw components that have
been logged per timestamp, across this entity and its recursive
children.
- `components`: instantiated once per `EntityTree` per `ComponentName`,
this keeps track of the number of raw components that have been logged
per timestamp per component, just for this entity.

For each of those we have to ask ourselves:
- Are they global scope? per entity? per component?
- Are they flat? recursive?
- Do they count the number of log events? of rows? the number of
distinct components? the number of raw components?!
- Do they ignore some components? Do they ignore auto-generated data?
- How do they behave in case of `Clear` events?
- Are they affected by GC or not? If so, in what way?
- Etc

Answers to these questions are now formalized in a test suite, A)
allowing us to easily iterate on them in the future and B) allowing me
to match the current behavior in new event-driven `EntityTree` that's
about to ship.

This also highlighted a bug in the current implementation, which is
fixed in the upcoming one.